requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
fakeredis==1.7.0
fastapi==0.71.0
fastapi-mail==0.4.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.5
httptools==0.3.0
httpx==0.21.3
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pydantic==1.9.0
pyparsing==3.0.6
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
redis==4.0.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wrapt==1.13.3
